divide app in django
