+++ classico_django/classico_board/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+# Register your models here.
