+++ classico_django/classico_board/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ClassicoBoardConfig(AppConfig):
+    name = 'classico_board'
