+++ classico_django/classico_keywords/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ClassicoKeywordsConfig(AppConfig):
+    name = 'classico_keywords'
