+++ classico_django/classico_keywords/urls.py
@@ -1,8 +1,8 @@
 from django.conf.urls import url, include
 from classico_app import views
 
-app_name = 'classico_app'
+app_name = 'classico_keywords'
 
 urlpatterns = [
-    url(r'^register/', views.register, name='register'),
+
 ]
