+++ classico_django/templates/classico_app/base/base.html
@@ -12,7 +12,7 @@
             <ul class="nav navbar-nav">
                 <li><a class="navbar-brand" href="{% url 'index' %}">Django</a></li>
                 <li><a class="navbar-link" href="{% url 'admin:index' %}">Admin</a></li>
-                <li><a class="navbar-link" href="{% url 'register' %}">Register</a></li>
+                <li><a class="navbar-link" href="{% url 'classico_app:register' %}">Register</a></li>
                 <li><a class="navbar-link" href="{% url 'rest-swagger' %}">REST-swagger</a></li>
