+++ classico_django/templates/classico_app/board/board_list.html
