+++ classico_django/templates/classico_app/index.html
@@ -1,4 +1,4 @@
-{% extends "classico_app/base.html" %}
+{% extends "classico_app/base/base.html" %}
 {% block body_block %}
     {% load staticfiles %}
     <div class="jumbotron">
