+++ classico_django/templates/classico_app/user/login.html
@@ -1,0 +1,15 @@
+{% extends 'classico_app/base.html' %}
+{% block body_block %}
+
+<div class="jumbotron"></div>
+    <h1>Please Login</h1>
+    <form action="{% url 'classico_app:user_login' %}" method="post">
+        {% csrf_token %}
+        <label for="username">User Name : </label>
+        <input type="text" name="username" placeholder="Enter Username">
+
+        <label for="password">Password : </label>
+        <input type="text" name="password" placeholder="Enter Password">
+
+        <input class="btn-u btn-u-green rounded" type="submit" name="login" value="Login">
+    </form>
