+++ classico_django/templates/classico_app/user/register.html
@@ -1,4 +1,4 @@
-{% extends 'classico_app/base.html' %}
+{% extends 'classico_app/base/base.html' %}
 {% load staticfiles %}
 
 {% block body_block %}
